The SI classes programmed with cowardly AI are now slayed instead of frozen when their aggressiveness expires
On a slightly delay, smokers are slayed when the "tongue_release" game
event fires, supposedly in all cases a survivor or dead survivor is
released from their tongue. Similarly spitters are slayed after firing
their spit as the aforementioned classes are programmed with AI to run
away after casting abilities. This appears to be done on a level too
deep for me to care or bother.